make smoke testing a real task
