+++ buildSrc/src/main/kotlin/dev/nolij/zumegradle/smoketest/SmokeTest.kt
@@ -1,0 +1,59 @@
+package dev.nolij.zumegradle.smoketest
+
+import org.gradle.api.logging.Logger
+import java.io.File
+import java.util.*
+import kotlin.io.path.*
+
+import dev.nolij.zumegradle.smoketest.Thread.ThreadState
+
+class SmokeTest(
+	internal val logger: Logger,
+	internal val portableMCBinary: String,
+	internal val modFile: File,
+	internal val mainDir: String,
+	internal val workDir: String,
+	internal val maxThreads: Int,
+	internal val threadTimeout: Long,
+	internal val configs: List<Config>
+) {
+	
+	private val threads = ArrayList<Thread>()
+	
+	internal fun printThreads() {
+		println("""
+				>   TOTAL: ${threads.filter { thread -> thread.finished }.size}/${configs.size}
+				> RUNNING: ${threads.filter { thread -> thread.state == ThreadState.RUNNING }.size}/${maxThreads}
+				>  PASSED: ${threads.filter { thread -> thread.state == ThreadState.PASSED }.size}
+				>  FAILED: ${threads.filter { thread -> thread.state == ThreadState.FAILED }.size}
+			""".trimIndent())
+	}
+	
+	fun test() {
+		println("Setting up instances...")
+		configs.forEach { config ->			
+			threads.add(Thread(this, config))
+		}
+
+		printThreads()
+
+		do {
+			while (threads.count { thread -> thread.isAlive } < maxThreads)
+				threads.firstOrNull { thread -> thread.state == ThreadState.PENDING }?.start() ?: break
+			java.lang.Thread.sleep(500L)
+			threads.forEach(Thread::update)
+		} while (threads.any { thread -> !thread.finished })
+
+		val failedConfigs = threads
+			.filter { thread -> thread.state == ThreadState.FAILED }
+			.map { thread -> thread.config }
+		
+		if (failedConfigs.isNotEmpty()) {
+			logger.error("[{\n${failedConfigs.joinToString("}, {\n")}}]")
+			error("One or more tests failed. See logs for more details.")
+		}
+
+		println("All tests passed.")
+	}
+	
+}
